Move test fixture classes to dedicated fixtures sub-packages
