+++ junit5/container/src/test/java/org/jboss/arquillian/junit5/container/fixtures/TestsWithIncludesExcludes.java
@@ -1,4 +1,4 @@
-package org.jboss.arquillian.junit5.container;
+package org.jboss.arquillian.junit5.container.fixtures;
 
 import org.jboss.arquillian.junit5.container.annotation.ArquillianTest;
 import org.junit.jupiter.api.Disabled;
